+++ src/app/indicator/sideIndicator.controller.js
@@ -12,10 +12,10 @@
     that.tree = $scope.dim.tree; // 耦合上下文
 
     // 删除一项指标, 指定code
-    that.remove = function(code) {
+    that.remove = function(id) {
       for (var i = 0, ilen = that.tree.childs.length; i < ilen; i++) {
         var node = that.tree.childs[i];
-        if (node.code === code) {
+        if (node.id === id) {
           return that.tree.childs.splice(i, 1);
         }
       }